token-metadata: fix test util organization to avoid circ refs
