+++ token-metadata/js/test/actions/index.ts
@@ -1,2 +1,3 @@
+export * from './create-collection';
 export * from './create-metadata';
 export * from './create-mint-account';
