+++ token-metadata/js/test/utils/address-labels.ts
@@ -1,5 +1,5 @@
 import { AddressLabels } from '@metaplex-foundation/amman';
-import { logDebug } from '.';
+import { logDebug } from './log';
 
 const persistLabelsPath = process.env.ADDRESS_LABEL_PATH;
 const knownLabels = { ['metaqbxxUerdq28cj1RbAWkYQm3ybzjb6a8bt518x1s']: 'mpl-token-metadata' };
