+++ token-metadata/js/test/utils/consts.ts
@@ -1,0 +1,12 @@
+import { LOCALHOST } from '@metaplex-foundation/amman';
+import { clusterApiUrl } from '@solana/web3.js';
+
+export const programIds = {
+  metadata: 'metaqbxxUerdq28cj1RbAWkYQm3ybzjb6a8bt518x1s',
+  vault: 'vau1zxA2LbssAUEF7Gpw91zMM1LvXrvpzJtmZ58rPsn',
+  auction: 'auctxRXPeJoc4817jDhf4HbjnhEcr1cCXenosMhK5R8',
+  metaplex: 'p1exdMJcjVao65QdewkaZRUnU6VPSXhus9n2GzWfh98',
+};
+
+export const DEVNET = clusterApiUrl('devnet');
+export const connectionURL = process.env.USE_DEVNET != null ? DEVNET : LOCALHOST;
