+++ token-metadata/js/test/utils/log.ts
@@ -1,0 +1,6 @@
+import debug from 'debug';
+
+export const logError = debug('mpl:tm-test:error');
+export const logInfo = debug('mpl:tm-test:info');
+export const logDebug = debug('mpl:tm-test:debug');
+export const logTrace = debug('mpl:tm-test:trace');
